目前有莫名其妙的错误

+++ zwj206/CLineTransDlg.h
@@ -12,8 +12,8 @@ public:
 	CLineTransDlg(CWnd* pParent = nullptr);   // 标准构造函数
 	virtual ~CLineTransDlg();
 
-	double m_param_a;
-	double m_param_b;
+	float m_param_a;
+	float m_param_b;
 
 // 对话框数据
 #ifdef AFX_DESIGN_TIME
